chore(react-ui): renamed index to textarea
