+++ libs/react-ui/src/index.ts
@@ -4,4 +4,4 @@ export {
   ButtonSize,
 } from './lib/button/button';
 
-export { default as TextareaComponent } from './lib/textarea/index';
+export { default as TextareaComponent } from './lib/textarea/textarea';
